+++ KrakEngine/Image.cpp
@@ -81,7 +81,7 @@ namespace KrakEngine{
             
             float timePerFrame = animation.Length / animation.NumFrames;
             UINT numxframes = (UINT)(g_GRAPHICSSYSTEM->GetBitmap(m_ImageName)->GetPixelSize().width / m_FrameWidth);
-            UINT currentFrame = (UINT)(m_Time / timePerFrame) + animation.StartFrame;
+            UINT currentFrame = (UINT)(m_Time / timePerFrame) + (UINT)animation.StartFrame;
 
             UINT currentXFrame = currentFrame % numxframes;
             UINT currentYFrame = currentFrame / numxframes;
